backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
requests
supabase

